Add global validator set, calc deposit data root on-chain, tweak interfaces

+++ contracts/interface/megapool/RocketMegapoolDelegateInterface.sol
@@ -17,7 +17,7 @@ interface RocketMegapoolDelegateInterface is RocketMegapoolDelegateBaseInterface
     function dissolveValidator(uint32 _validatorId) external;
     function getNodeAddress() external returns (address);
 
-    function getValidatorCount() external view returns (uint256);
+    function getValidatorCount() external view returns (uint32);
     function getValidatorInfo(uint32 _validatorId) external view returns (RocketMegapoolStorageLayout.ValidatorInfo memory);
     function getAssignedValue() external view returns (uint256);
     function getDebt() external view returns (uint256);
